+++ src/app/modules/Folder/folder.service.ts
@@ -1,3 +1,4 @@
+import AppError from "../../error/AppError";
 import { TFolder } from "./folder.interface";
 import { FolderModel } from "./folder.model";
 
@@ -6,17 +7,40 @@ const createFolderIntoDB = async (payload: TFolder) => {
   return result;
 };
 
-const getAllFolderFromDB = async (userId: string) => {
-  const result = await FolderModel.find({userId});
-  return result;
+const getAllFolderFromDB = async (accountId: string) => {
+  const result = await FolderModel.find({ userId: accountId });
+
+  const totalItem = result.length;
+
+  const jsonData = JSON.stringify(result);
+  const sizeInBytes = Buffer.byteLength(jsonData, "utf8");
+  const usagesStorageInGB = sizeInBytes / 1e9;
+
+  return {
+    datas: result,
+    metaData: {
+      totalItem,
+      usagesStorageInGB,
+    },
+  };
 };
 
 const getSingleFolderFromDB = async (id: string) => {
+  const isFolderExists = await FolderModel.findById(id);
+  if (!isFolderExists) {
+    throw new AppError(404, "Folder dose not exists");
+  }
+
   const result = await FolderModel.findById(id);
   return result;
 };
 
 const updateFolderFromDB = async (id: string, payload: { name: string }) => {
+  const isFolderExists = await FolderModel.findById(id);
+  if (!isFolderExists) {
+    throw new AppError(404, "Folder dose not exists");
+  }
+
   const result = await FolderModel.findByIdAndUpdate(id, payload, {
     new: true,
     runValidators: true,
@@ -25,6 +49,11 @@ const updateFolderFromDB = async (id: string, payload: { name: string }) => {
 };
 
 const makeFavoritFolderIntoDB = async (id: string) => {
+  const isFolderExists = await FolderModel.findById(id);
+  if (!isFolderExists) {
+    throw new AppError(404, "Folder dose not exists");
+  }
+
   const result = await FolderModel.findByIdAndUpdate(
     id,
     { isFavorite: true },
@@ -37,6 +66,11 @@ const makeFavoritFolderIntoDB = async (id: string) => {
 };
 
 const deleteFolderFromDB = async (id: string) => {
+  const isFolderExists = await FolderModel.findById(id);
+  if (!isFolderExists) {
+    throw new AppError(404, "Folder dose not exists");
+  }
+
   const result = await FolderModel.findByIdAndDelete(id);
   return result;
 };
